carbon-calculator/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
anthropic==0.25.0
python-dotenv==1.0.0
gunicorn==21.2.0